requirements.txt
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
Flask==2.2.5
gunicorn==23.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.21.6
packaging==24.0
pickle-mixin==1.0.2
scikit-learn==1.0.2
scipy==1.7.3
python-dotenv==0.21.0
threadpoolctl==3.1.0
typing_extensions==4.7.1
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
